--- a/lib/loaders/guui-css-loader.js
+++ b/lib/loaders/guui-css-loader.js
@@ -2,11 +2,6 @@
 
 const { parseCSS: emotionParser } = require('babel-plugin-emotion/lib/parser');
 
-type CategorisedStyles = {
-    expensiveCss?: Object,
-    cheapCss?: Object,
-};
-
 // turn class selectors into reasonable keys
 // {'.header': 'blah'} => {'header': 'blah'}
 const normaliseKeys = (styles: Object): Object =>
@@ -34,58 +29,6 @@ const reFloat = (rules: Object): Object =>
         return Object.assign({}, refloated, { [key]: rules[key] });
     }, {});
 
-// splits rules into cheap/expensive clusters which
-// can then be handled appropriately. which
-// basically means 'can styletron handle it?':
-// {
-//     color: 'red',
-//     '@supports (display: flex)': {
-//         display: 'flex',
-//     },
-// }
-// becomes
-// {
-//     cheapCSS: { color: 'red' },
-//     expensiveCSS: {
-//         '@supports (display: flex)': {
-//             display: 'flex',
-//         },
-//     },
-// }
-const categoriseStyles = (styles: Object): CategorisedStyles =>
-    Object.keys(styles).reduce((categorisedStyles, key) => {
-        const rule = styles[key];
-
-        const categorisedStyle = Object.keys(
-            rule
-        ).reduce(({ cheapCSS = {}, expensiveCSS = {} }, decl) => {
-            if (
-                typeof rule[decl] !== 'object' ||
-                (decl.startsWith('@media') &&
-                    Object.values(rule[decl]).every(
-                        mediaDeclValue => typeof mediaDeclValue !== 'object'
-                    ))
-            ) {
-                return {
-                    expensiveCSS,
-                    cheapCSS: Object.assign({}, cheapCSS, {
-                        [decl]: rule[decl],
-                    }),
-                };
-            }
-            return {
-                cheapCSS,
-                expensiveCSS: Object.assign({}, expensiveCSS, {
-                    [decl]: rule[decl],
-                }),
-            };
-        }, {});
-
-        return Object.assign(categorisedStyles, {
-            [key]: categorisedStyle,
-        });
-    }, {});
-
 module.exports = function uiCSS(source: string): string {
     const { styles } = reFloat(
         // have emotion parse it
@@ -95,8 +38,5 @@ module.exports = function uiCSS(source: string): string {
     // turn '.key' keys into 'key'
     const normalisedKeysStyles = normaliseKeys(styles);
 
-    // split out styles we can pass to styeltron
-    const categorisedStyles = categoriseStyles(normalisedKeysStyles);
-
-    return `module.exports = ${JSON.stringify(categorisedStyles)};`;
+    return `module.exports = ${JSON.stringify(normalisedKeysStyles)};`;
 };