--- a/src/extractCriticalCss.js
+++ b/src/extractCriticalCss.js
@@ -2,6 +2,6 @@
 import { extractCritical } from 'emotion-server';
 
 const extractCriticalCss = (body: string): string =>
-    `<style expensive-css>${extractCritical(body).css}</style>`;
+    `<style>${extractCritical(body).css}</style>`;
 
 export default extractCriticalCss;