--- a/src/extractCriticalCss.js
+++ b/src/extractCriticalCss.js
@@ -1,7 +1,0 @@
-// @flow
-import { extractCritical } from 'emotion-server';
-
-const extractCriticalCss = (body: string): string =>
-    `<style>${extractCritical(body).css}</style>`;
-
-export default extractCriticalCss;