--- a/src/h.jsx
+++ b/src/h.jsx
@@ -1,8 +1,7 @@
 // @flow
 /* eslint-disable camelcase */
 
-// bespoke wrapper around preact's `h` that passes any CSS data
-// on `attributes.style` to styletron, then hands off to `preact#h`
+// bespoke wrapper around preact's `h`
 
 import { h as preact_h } from 'preact';
 import { css } from 'emotion';