--- a/src/server.js
+++ b/src/server.js
@@ -1,17 +1,14 @@
 // @flow
-import StyletronServer from 'styletron-server';
 import renderToString from './renderToString';
 import extractCriticalCss from './extractCriticalCss';
 
 const startServer = () => {
-    const server = new StyletronServer();
-
     return {
         renderToString(node: React.Element<any>): string {
-            return renderToString(server, node);
+            return renderToString(node);
         },
         extractCriticalCss(body: string): string {
-            return extractCriticalCss(server, body);
+            return extractCriticalCss(body);
         },
     };
 };