--- a/src/server.js
+++ b/src/server.js
@@ -1,11 +1,11 @@
 // @flow
-import renderToString from './renderToString';
+import { render as renderToStringPreact } from 'preact-render-to-string';
 import extractCriticalCss from './extractCriticalCss';
 
 const startServer = () => {
     return {
         renderToString(node: React.Element<any>): string {
-            return renderToString(node);
+            return renderToStringPreact(node);
         },
         extractCriticalCss(body: string): string {
             return extractCriticalCss(body);