--- a/src/server.js
+++ b/src/server.js
@@ -1,16 +1,19 @@
 // @flow
 import { render as renderToStringPreact } from 'preact-render-to-string';
-import extractCriticalCss from './extractCriticalCss';
+import { extractCritical } from 'emotion-server';
 
-const startServer = () => {
-    return {
-        renderToString(node: React.Element<any>): string {
-            return renderToStringPreact(node);
-        },
-        extractCriticalCss(body: string): string {
-            return extractCriticalCss(body);
-        },
-    };
+type GuuiServer = {
+    renderToString: (React.Element<any>) => string,
+    extractCriticalCss: string => string,
 };
+
+const startServer = (): GuuiServer => ({
+    renderToString(node: React.Element<any>): string {
+        return renderToStringPreact(node);
+    },
+    extractCriticalCss(body: string): string {
+        return `<style>${extractCritical(body).css}</style>`;
+    },
+});
 
 export default startServer;