(Somewhat limited and inconclusive) benchmarks on 404 page:

## Lighthouse (mean of 5 runs)

```
                            with Styletron  without Styletron
First meaningful paint (ms)       710              692
First interactive (ms)           2168             2300
Consistently interactive (ms)    2168             2300
Perceptual speed index           1219             1406
Estimated input latency (ms)     16.2             16.6

```

## Page weight

with Styletron: 34.3KB
without Styletron: 34.3KB

## HTML parse times (x6 slowdown, mean of 5 runs)

```
                            with Styletron  without Styletron
Initial parse (ms)                298              290
Parse after JS arrives (ms)       107               93
```